Finish exercise 05-flex-modal of the flex folder
Modify the index.html file to wrap the icon and information in
different divs. Within the information container, wrap the title and
close button elements, the text and the continue and cancel buttons
in different divs. Create different classes for each.

Modify the style.css file to position the elements of the index.html
file using flexbox.